+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## Unreleased changes
 
+- Add V1 variant of root and level 1 key updates.
 - Expose helper methods to wait for new blocks or finalization.
+- Add some helper methods to `InstanceInfo`, `ElectionDifficulty` and
+  `DelegationTarget` types.
 
 ## 1.0.2